Make the wsgi module a true WSGI endpoint.
